HSV filtering and first GUI sliders for it Simulation mode for non-raspi tests

+++ ShelfCam.py
@@ -2,7 +2,6 @@
 import cv2
 import sys
 import time
-import RPi.GPIO as GPIO
 import datetime
 import getopt
 #from takepic import *
@@ -13,25 +12,24 @@ import numpy as np
 from loadsettings import *
 from GUI import *
 from contour_operations import ContourOperations
-from pivideostream import PiVideoStream
+
 
 def usage():
-	print("usage: -g gui <1/0>")
+	print("usage: -g gui <1/0> -s sim <1/0>")
+	#simulation mode is for non- raspberry pi computers...
 
 def guiThread():
 	root = Tk()
-	root.geometry("800x270+0+0")
+	root.geometry("1000x1000+0+0")
 	app = Window(root)
 	root.mainloop()
 
 def main(argv):	
-	vs = PiVideoStream((1296, 736)).start()
-	#vs = PiVideoStream((1920, 1088)).start()
-	time.sleep(1)
 	#todo: crop xy threshold, pipette, mask, contours 
 	use_gui=0
+	simulate=0
 	try: 
-			opts, args =getopt.getopt(sys.argv[1:], "g:",["gui="])
+			opts, args =getopt.getopt(sys.argv[1:], "g:s:",["gui=","sim="])
 	except getopt.GetoptError as err:
 		usage()
 		sys.exit(2)
@@ -39,9 +37,12 @@ def main(argv):
 	#print("args: ",sys.argv[1:])
 	if len(opts)<1:
 		use_gui=0
+		simulate=0
 	for o,a, in opts:
 		if o=="-g":
 			use_gui=int(a)
+		elif o=="-s":
+			simulate=int(a)
 		
 	t_gui=Thread(target=guiThread)
 	if use_gui>0: 
@@ -50,19 +51,34 @@ def main(argv):
 		load_settings()
 		#print("set guiCommands nogui:",guiCommands)
 		guiCommands['previewRaw']=False
+	if simulate==0:
+		import RPi.GPIO as GPIO
+		from pivideostream import PiVideoStream
+		vs = PiVideoStream((1296, 736)).start()
+		#vs = PiVideoStream((1920, 1088)).start()
+		time.sleep(1)
 	runVideo=True
 	imageCounter=0
 	#print("picSavePath: ",picSavePath)
 	while (runVideo==True):
 		#print("mainloop")
 		runVideo=guiCommands['runVideo']
-		grabbedFrame = vs.readCropped(guiCommands['cropleft'],guiCommands['croptop'],guiCommands['cropright'],guiCommands['cropbottom'])
-		if guiCommands['takePic']==True:
+		if simulate==0:
+			grabbedFrame = vs.readCropped(guiCommands['cropleft'],guiCommands['croptop'],guiCommands['cropright'],guiCommands['cropbottom'])
+		else:
+			grabbedFrame=cv2.imread("images/image_000.png", -1)
+			#runVideo=False
+		if guiCommands['takePic']==True and simulate==0:
 			filename="images/image_"+str(imageCounter).zfill(3)+".png"
 			cv2.imwrite(filename, grabbedFrame)
 			imageCounter+=1
 			guiCommands['takePic']=False
 		Live=ContourOperations()
+		frame_hsv= cv2.cvtColor(grabbedFrame,cv2.COLOR_BGR2HSV)
+		lower_red=np.array([50,0,0])
+		upper_red=np.array([255,255,255])
+		mask=cv2.inRange(frame_hsv, lower_red, upper_red)
+		filtered= cv2.bitwise_and(grabbedFrame, grabbedFrame, mask=mask)
 		blue, green, red = cv2.split(grabbedFrame)
 		background=cv2.imread("background/image_000.png", -1)
 		bluebg, greenbg, redbg = cv2.split(grabbedFrame)
@@ -76,6 +92,8 @@ def main(argv):
 		
 		if guiCommands['previewRaw']==True: 
 			cv2.imshow('All',grabbedFrame)
+			cv2.imshow('Mask',mask)
+			cv2.imshow('Filtered',filtered)
 			#cv2.imshow('All-BG',cv2.subtract(grabbedFrame,background))
 			Live.showContour(grabbedFrame, contour_ext)
 			cv2.imshow('red',red_Threshold)
